UXODS-97 Menu Bar [Code] - separated out the menu component
