+++ src/js/design-system.ts
@@ -1,6 +1,7 @@
 import { AccordionBehavior } from './components/accordion/accordion';
 import { ModalDialogBehavior } from "./components/modal-dialog/modal-dialog";
 import {HeaderBehavior} from "./components/header/header";
+import {MenuToggleBehavior} from "./components/menu-toggle/menu-toggle";
 
 export class PrincetonDesignSystem {
     public enableDesignSystem() {
@@ -12,11 +13,13 @@ export class PrincetonDesignSystem {
         new AccordionBehavior().enable();
         new ModalDialogBehavior().enable();
         new HeaderBehavior().enable();
+        new MenuToggleBehavior().enable();
     }
 
     public disable() {
         new AccordionBehavior().disable();
         new ModalDialogBehavior().disable();
         new HeaderBehavior().disable();
+        new MenuToggleBehavior().disable();
     }
 }
