+++ stories/MenuToggle.stories.mdx
@@ -4,22 +4,103 @@ import { html } from 'lit-html';
 <Meta title="Components/Menu Toggle" component="MenuToggle" />
 
 export const Template = (args) => html`
-    <button class="emc-menu-toggle" type="button" aria-controls="menu-container" aria-expanded="false" aria-label="Toggle navigation">${ args.title }</button>
-`;
-
-export const TemplateExpanded = (args) => html`
-    <button class="emc-menu-toggle" type="button" aria-controls="menu-container" aria-expanded="true" aria-label="Toggle navigation">${ args.title }</button>
-`;
-
-export const TemplateWebComponent = (args) => html`
-    <emc-menu-toggle></emc-menu-toggle>
-`;
-
-export const TemplateChild = (args) => html`
-    <emc-child-test>
-        <emc-child name="Test1"></emc-child>
-        <emc-child name="Test2"></emc-child>
-    </emc-child-test>
+<!--<header role="banner" class="emc-header">-->
+    <nav aria-label="Main Menu" class="emc-menu emc-menu__main-menu">
+        <h2 class="emc-sr-only">Main Menu</h2>
+        <div class="emc-container">
+        <div class="emc-menu__main-menu-navbar">
+            <button class="emc-menu__menu-toggle" aria-expanded="false" aria-label="Navigation Menu Toggle">Menu<i class="emc-icon emc-icon-menu" aria-hidden="true"></i></button>
+            <div class="emc-menu__nav-container">
+                <ul class="emc-menubar" role="list">
+                    <li aria-current="true">
+                        <a href="javascript:void(0);" aria-current="page">Content Types</a>
+                        <button class="emc-menu__submenu-toggle" type="button" aria-expanded="${ args.forceExpanded }">
+                            <span class="emc-sr-only">
+                                Content Types
+                                SubMenu
+                            </span>
+                        </button>
+                        <ul>
+                            <li>
+                                <a href="javascript:void(0);">Page</a>
+                                <button type="button" class="emc-menu__submenu-toggle" aria-expanded="${ args.forceExpanded }">
+                                    <span class="emc-sr-only">
+                                        Page
+                                        SubMenu
+                                    </span>
+                                </button>
+                                <ul class="emc-menubar__submenu--shown-by-default">
+                                    <li><a href="javascript:void(0);">Page List</a></li>
+                                    <li><a href="javascript:void(0);" aria-current="page">Sub Page</a></li>
+                                    <li><a href="javascript:void(0);">Sub Page with Right Sidebar</a></li>
+                                    <li><a href="javascript:void(0);">Landing Page</a></li>
+                                    <li><a href="javascript:void(0);">Detail Page</a></li>
+                                </ul>
+                             </li>
+                            <li aria-current="page"><a href="javascript:void(0);">Alert</a></li>
+                            <li>
+                                <a href="javascript:void(0);">News</a>
+                                <button type="button" aria-expanded="false">
+                                    <span class="emc-sr-only">
+                                        News
+                                        SubMenu
+                                    </span>
+                                </button>
+                                <ul>
+                                    <li><a href="javascript:void(0);">News List</a></li>
+                                    <li><a href="javascript:void(0);">News List with Archive Years</a></li>
+                                    <li><a href="javascript:void(0);">News Landing Page</a></li>
+                                    <li><a href="javascript:void(0);">Featured News</a></li>
+                                </ul>
+                            </li>
+                            <li><a href="javascript:void(0);">Person</a></li>
+                        </ul>
+                    </li>
+                    <li>
+                        <a href="javascript:void(0);">Blocks</a>
+                        <button type="button" class="emc-menu__submenu-toggle" aria-expanded="false">
+                            <span class="emc-sr-only">
+                                Blocks
+                                SubMenu
+                            </span>
+                        </button>
+                        <ul>
+                            <li><a href="javascript:void(0);">Accordion</a></li>
+                            <li><a href="javascript:void(0);">Content Slider</a></li>
+                            <li><a href="javascript:void(0);">Feature</a></li>
+                            <li><a href="javascript:void(0);">iFrame</a></li>
+                            <li><a href="javascript:void(0);">Map</a></li>
+                            <li><a href="javascript:void(0);">Related Links</a></li>
+                        </ul>
+                    </li>
+                    <li>
+                        <a href="javascript:void(0);">Images &amp; Media</a>
+                        <button class="emc-menu__submenu-toggle" type="button" aria-expanded="false">
+                            <span class="emc-sr-only">
+                                Images &amp; Media
+                                SubMenu
+                            </span>
+                        </button>
+                        <ul>
+                            <li><a href="javascript:void(0);">Featured Images</a></li>
+                            <li><a href="javascript:void(0);">Billboard Image</a></li>
+                            <li><a href="javascript:void(0);">Billboard Slider</a></li>
+                            <li><a href="javascript:void(0);">Billboard Video</a></li>
+                            <li><a href="javascript:void(0);">Image Gallery</a></li>
+                            <li><a href="javascript:void(0);">Image Embed</a></li>
+                            <li><a href="javascript:void(0);">Social Media</a></li>
+                        </ul>
+                    </li>
+                </ul>
+            </div>
+        </div>
+        </div>
+    </nav>
+<!--</header>-->
+    <a name="main-content"></a>
+    <script id="initscript">
+        PrincetonDesignSystem.enableDesignSystem();
+    </script>
 `;
 
 # Menu Toggle
@@ -32,32 +113,10 @@ The Menu Toggle is a common pattern for toggling the visibility of a menu.
 
 <Canvas>
     <Story name="Collapsed" args={{
-        title: "Menu"
+        forceExpanded: false,
+        siteBrandingName: 'Relativity',
+        siteBrandingSlogan: 'The Princeton University Design System'
     }}>
       {Template.bind({})}
-    </Story>
-</Canvas>
-
-<Canvas>
-    <Story name="Expanded" args={{
-        title: "Menu"
-    }}>
-      {TemplateExpanded.bind({})}
-    </Story>
-</Canvas>
-
-<Canvas>
-    <Story name="Web Component" args={{
-        title: "Menu"
-    }}>
-      {TemplateWebComponent.bind({})}
-    </Story>
-</Canvas>
-
-<Canvas>
-    <Story name="Child" args={{
-        title: "Menu"
-    }}>
-      {TemplateChild.bind({})}
     </Story>
 </Canvas>
